% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merMod.R
\name{dataClasses.merMod}
\alias{dataClasses.merMod}
\title{\link[ecip]{dataClasses} for \code{merMod}}
\usage{
\method{dataClasses}{merMod}(x)
}
\arguments{
\item{x}{..

Function \code{lme4:::terms.merMod()} return does not have \code{'dataClasses'} attribute (as of 2025-03-25).}
}
\description{
\link[ecip]{dataClasses} for \code{merMod}
}
\keyword{internal}
