% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0PACKAGE.R
\docType{package}
\name{lme4.tzh-package}
\alias{lme4.tzh}
\alias{lme4.tzh-package}
\title{lme4.tzh: Tingting's extension of lme4 package}
\description{
Tingting's extension of lme4 package.
}
\author{
\strong{Maintainer}: Tingting Zhan \email{tingtingzhan@gmail.com} (\href{https://orcid.org/0000-0001-9971-4844}{ORCID})

}
\keyword{internal}
