% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merMod.R
\name{md_.merMod}
\alias{md_.merMod}
\title{R Markdown Lines for \code{merMod}}
\usage{
\method{md_}{merMod}(x, xnm, ...)
}
\arguments{
\item{x, xnm, ...}{..}
}
\description{
R Markdown Lines for \code{merMod}
}
\examples{
library(ecip) 
list(
 '`glmerMod`' = glmer(cbind(incidence, size-incidence) ~ period + (1|herd), 
     data = cbpp, family = binomial)
) |> rmd.tzh::render_(file = 'glmerMod')
  
}
\keyword{internal}
