% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merMod.R
\name{rmd_.merMod}
\alias{rmd_.merMod}
\title{rmd_.merMod}
\usage{
rmd_.merMod(x, xnm, autofold = TRUE, ...)
}
\arguments{
\item{x}{a \link[lme4]{merMod} object}

\item{xnm}{..}

\item{autofold}{\link[base]{logical} scalar}

\item{...}{..}
}
\description{
rmd_.merMod
}
