% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlmerMod.R
\name{s3_rlmerMod}
\alias{s3_rlmerMod}
\alias{desc_.rlmerMod}
\alias{coef_.rlmerMod}
\alias{confint.rlmerMod}
\alias{.pval.summary.rlmerMod}
\alias{ngrps.rlmerMod}
\alias{nobsText.rlmerMod}
\title{Additional S3 Method Dispatches for \link[robustlmm]{rlmer} Return}
\usage{
\method{desc_}{rlmerMod}(x)

\method{coef_}{rlmerMod}(x)

\method{confint}{rlmerMod}(object, parm, level = 0.95, ...)

\method{.pval}{summary.rlmerMod}(x)

\method{ngrps}{rlmerMod}(object, ...)

\method{nobsText}{rlmerMod}(x)
}
\arguments{
\item{x, object}{a \link[lme4]{merMod} object}

\item{parm, level}{see \link[stats]{confint.default}}

\item{...}{additional parameters, currently not in use}
}
\description{
..
}
\details{
Package \CRANpkg{robustlmm} imports package \CRANpkg{lme4}.

Object returned by function \link[robustlmm]{rlmer} does \strong{not} \link[base]{inherits} from \link[lme4]{merMod} class.
}
\examples{
# ?robustlmm::rlmer
library(robustlmm)
m = rlmer(Reaction ~ Days + (Days|Subject), data = sleepstudy,
 rho.sigma.e = psi2propII(smoothPsi, k = 2.28),
 rho.b = chgDefaults(smoothPsi, k = 5.14, s=10),
 rho.sigma.b = chgDefaults(smoothPsi, k = 5.14, s=10))

ecip::coef_(m)
confint(m)
}
