% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merMod.R, R/rlmerMod.R
\name{s3_merMod}
\alias{s3_merMod}
\alias{.pval.summary.merMod}
\alias{vterms.merMod}
\alias{desc_.glmerMod}
\alias{desc_.merMod}
\alias{coef_.merMod}
\alias{confint_.merMod}
\alias{nobsText.merMod}
\alias{coef_.rlmerMod}
\title{S3 Method Dispatches for \link[lme4]{merMod} Class}
\usage{
\method{.pval}{summary.merMod}(x)

\method{vterms}{merMod}(x)

\method{desc_}{glmerMod}(x)

\method{desc_}{merMod}(x)

\method{coef_}{merMod}(x)

\method{confint_}{merMod}(x, level = 0.95, method = "Wald", ...)

\method{nobsText}{merMod}(x)

\method{coef_}{rlmerMod}(x)
}
\arguments{
\item{x}{a \link[lme4]{merMod} object}

\item{level}{\link[base]{double} scalar, confidence level}

\item{method}{..}

\item{...}{..}
}
\description{
..
}
\examples{
# see ?lme4::glmer or ?lme4::cbpp
library(lme4)
class(m1 <- glmer(cbind(incidence, size - incidence) ~ period + (1 | herd), 
  data = cbpp, family = binomial))
desc_.glmerMod(m1)
coef_.merMod(m1)
confint_.merMod(m1)
nobsText.merMod(m1)

startvec = c(Asym = 200, xmid = 725, scal = 350)
(m2 <- nlmer(circumference ~ SSlogis(age, Asym, xmid, scal) ~ Asym|Tree,
  Orange, start = startvec))
class(m2)
desc_.merMod(m2)


library(rmd.tzh); library(ecip) 
list(
 '`glmerMod`' = glmer(cbind(incidence, size-incidence) ~ period + (1|herd), 
     data = cbpp, family = binomial)
) |> render_(file = 'glmerMod')
  
library(lme4)
library(HSAUR3)
m = glmer(outcome ~ treatment + visit + (1|patientID), data=toenail, family=binomial, nAGQ=20)
m |> vterms.merMod()
}
\keyword{internal}
