% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merMod.R
\name{s3_merMod}
\alias{s3_merMod}
\alias{.pval.merMod}
\alias{model_desc.glmerMod}
\alias{model_desc.merMod}
\alias{coef0.merMod}
\alias{confint2.merMod}
\alias{nobsText.merMod}
\title{Additional S3 Method Dispatches for \link[lme4]{merMod} Class}
\usage{
.pval.merMod(x)

model_desc.glmerMod(x, ...)

model_desc.merMod(x, ...)

coef0.merMod(x)

confint2.merMod(object, method = "Wald", ...)

nobsText.merMod(x)
}
\arguments{
\item{x, object}{a \link[lme4]{merMod} object}

\item{...}{..}

\item{method}{..}
}
\description{
..
}
\examples{
# see ?lme4::glmer or ?lme4::cbpp
library(lme4)
class(m1 <- glmer(cbind(incidence, size - incidence) ~ period + (1 | herd), 
  data = cbpp, family = binomial))
.pval.merMod(m1)

}
