% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merMod.R, R/rlmerMod.R
\name{s3_merMod}
\alias{s3_merMod}
\alias{.pval.merMod}
\alias{.pval.summary.merMod}
\alias{desc_.glmerMod}
\alias{desc_.merMod}
\alias{coef0.merMod}
\alias{confint_.merMod}
\alias{nobsText.merMod}
\alias{coef0.rlmerMod}
\title{S3 Method Dispatches for \link[lme4]{merMod} Class}
\usage{
.pval.merMod(x)

.pval.summary.merMod(x)

desc_.glmerMod(x)

desc_.merMod(x)

coef0.merMod(x)

confint_.merMod(object, method = "Wald", ...)

nobsText.merMod(x)

coef0.rlmerMod(x)
}
\arguments{
\item{x, object}{a \link[lme4]{merMod} object}

\item{method}{..}

\item{...}{..}
}
\description{
..
}
\examples{
# see ?lme4::glmer or ?lme4::cbpp
library(lme4)
class(m1 <- glmer(cbind(incidence, size - incidence) ~ period + (1 | herd), 
  data = cbpp, family = binomial))
.pval.merMod(m1)
desc_.glmerMod(m1)
coef0.merMod(m1)
confint_.merMod(m1)
nobsText.merMod(m1)

startvec = c(Asym = 200, xmid = 725, scal = 350)
(m2 <- nlmer(circumference ~ SSlogis(age, Asym, xmid, scal) ~ Asym|Tree,
  Orange, start = startvec))
class(m2)
}
